refactor(memory): plugin-owned memory-db accessor; drop dead v3 edge tables
Follow-up to #37798:

- The fail-soft memory-connection accessor moves out of injection-events.ts
  into plugins/defaults/memory/memory-db.ts as the single place plugin code
  resolves assistant-memory.db — upcoming table relocations all consume it.
- Migration 328 drops memory_v3_coactivation and memory_v3_auto_edges: the
  v3 edge-learning subsystem that produced and consumed them was removed
  (memory_v3_edge_learning sits in the jobs-worker retired-type drop list),
  so the tables only grew the main DB with rows nothing reads.
- conversation_starters moves from schema/memory-core.ts to its own
  schema/conversation-starters.ts — it's home-surface product state, not
  memory-plugin storage; consumers import via the schema index and are
  unchanged.
