+++ assistant/src/persistence/schema/index.ts
@@ -4,6 +4,7 @@ export * from "./bookmarks.js";
 export * from "./calls.js";
 export * from "./contacts.js";
 export * from "./conversation-groups.js";
+export * from "./conversation-starters.js";
 export * from "./conversations.js";
 export * from "./documents.js";
 export * from "./guardian.js";
